fixed problem with max

+++ Kinetic_data_structures/include/CGAL/Kinetic/Delaunay_triangulation_2.h
@@ -753,7 +753,8 @@ protected:
     return vhs_[k.to_index()];
   }
   void set_vertex_handle(Point_key k, Vertex_handle vh) {
-    vhs_.resize(std::max(k.to_index()+1, vhs_.size()));
+    vhs_.resize(std::max(static_cast<unsigned int>(k.to_index()+1),
+			 static_cast<unsigned int>(vhs_.size())));
     vhs_[k.to_index()]=vh;
   }
